Extract functions to init_views module
